+++ SemiPlot/SemiPlot.Tests/Chart/DeltaCursorModelTests.cs
@@ -127,7 +127,7 @@ public sealed class DeltaCursorModelTests
 		model.HasBothCursors.Should().BeFalse();
 	}
 
-	private static PenHistoryEnvelope Envelope(long penId, params (DateTime Time, double Center)[] columns)
+	private static PenHistoryEnvelope Envelope(int penId, params (DateTime Time, double Center)[] columns)
 	{
 		var timestamps = new DateTime[columns.Length];
 		var min = new double[columns.Length];
